ai-education-service/requirements.txt
# AI 教育资料处理服务依赖

# Web 框架
fastapi
uvicorn[standard]

# 配置管理
pydantic
pydantic-settings

# 阿里云 OSS
oss2

# 阿里云 DashVector
dashvector

# 阿里云 DashScope
dashscope

# LlamaIndex 核心
llama-index-core

# LlamaIndex Workflows 可视化工具
llama-index-utils-workflow

# LlamaIndex 文件读取器
llama-index-readers-file

# LlamaParse
llama-cloud-services

# HTTP 客户端
httpx

# 重试机制
tenacity

# 文档处理
pypdf
python-docx
python-pptx
docx2txt

# 日志
rich

# Neo4j 知识图谱
neo4j

# PostgreSQL 数据库（用户记忆存储）
psycopg2-binary

# LangGraph 多智能体
langgraph
langchain-core
